+++ src/init.c
@@ -25,7 +25,7 @@ extern SEXP _ClusterR_opt_clust_fK(SEXP, SEXP, SEXP);
 extern SEXP _ClusterR_predict_medoids(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _ClusterR_predict_MGausDPDF(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _ClusterR_Predict_mini_batch_kmeans(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _ClusterR_preferenceRange(SEXP, SEXP);
+extern SEXP _ClusterR_preferenceRange(SEXP, SEXP, SEXP);
 extern SEXP _ClusterR_SCALE(SEXP, SEXP, SEXP);
 extern SEXP _ClusterR_split_rcpp_lst(SEXP);
 extern SEXP _ClusterR_validate_centroids(SEXP, SEXP);
@@ -48,7 +48,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_ClusterR_predict_medoids",           (DL_FUNC) &_ClusterR_predict_medoids,            7},
     {"_ClusterR_predict_MGausDPDF",         (DL_FUNC) &_ClusterR_predict_MGausDPDF,          5},
     {"_ClusterR_Predict_mini_batch_kmeans", (DL_FUNC) &_ClusterR_Predict_mini_batch_kmeans,  4},
-    {"_ClusterR_preferenceRange",           (DL_FUNC) &_ClusterR_preferenceRange,            2},
+    {"_ClusterR_preferenceRange",           (DL_FUNC) &_ClusterR_preferenceRange,            3},
     {"_ClusterR_SCALE",                     (DL_FUNC) &_ClusterR_SCALE,                      3},
     {"_ClusterR_split_rcpp_lst",            (DL_FUNC) &_ClusterR_split_rcpp_lst,             1},
     {"_ClusterR_validate_centroids",        (DL_FUNC) &_ClusterR_validate_centroids,         2},